tech-news/requirements.txt
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
PyYAML==6.0.1
python-dateutil==2.8.2
google-generativeai==0.3.2
pytest==7.4.3
pytest-mock==3.12.0
